Allow pleb's to edit/delete personal categories

+++ src/pages/api/cats/index.ts
@@ -3,13 +3,19 @@ import { getCategories } from '@/prisma/queries'
 import dayjs from 'dayjs'
 import type { NextApiRequest, NextApiResponse } from 'next'
 import { Entry } from '@prisma/client'
+import { getToken } from 'next-auth/jwt'
 
 const handler = async (req: NextApiRequest, res: NextApiResponse) => {
   switch (req.method) {
     case 'GET':
       const categories = await getCategories()
       return res.status(200).json(categories)
-    case 'PUT':
+    case 'PUT': {
+      const token = await getToken({ req })
+      // Only admins can edit master categories
+      if (req.body.isMaster && !token?.isAdmin) {
+        return res.status(401).send('Unauthorized')
+      }
       try {
         const [, editedCategory] = await prisma.$transaction([
           prisma.entry.deleteMany({
@@ -34,9 +40,9 @@ const handler = async (req: NextApiRequest, res: NextApiResponse) => {
               entries: {
                 createMany: {
                   data: req.body.dates.map(
-                    ({ date, isRepeating = false }: { date: string; isRepeating?: boolean }) => ({
+                    ({ date, isRecurring = false }: { date: string; isRecurring?: boolean }) => ({
                       date: dayjs(date).toISOString(),
-                      isRepeating: isRepeating
+                      isRecurring: isRecurring
                     })
                   ),
                   skipDuplicates: true
@@ -52,7 +58,13 @@ const handler = async (req: NextApiRequest, res: NextApiResponse) => {
       } catch (error) {
         return res.status(404).send('category does not exist')
       }
-    case 'POST':
+    }
+    case 'POST': {
+      const token = await getToken({ req })
+      // Only admins can create master categories
+      if (req.body.isMaster && !token?.isAdmin) {
+        return res.status(401).send('Unauthorized')
+      }
       const new_category = await prisma.category.create({
         data: {
           name: req.body.name,
@@ -66,9 +78,9 @@ const handler = async (req: NextApiRequest, res: NextApiResponse) => {
           endDate: req.body.endDate ? dayjs(req.body.endDate).toISOString() : null,
           entries: {
             createMany: {
-              data: req.body.dates.map(({ date, isRepeating = false }: { date: string; isRepeating?: boolean }) => ({
+              data: req.body.dates.map(({ date, isRecurring = false }: { date: string; isRecurring?: boolean }) => ({
                 date: dayjs(date).toISOString(),
-                isRepeating: isRepeating
+                isRecurring: isRecurring
               })),
               skipDuplicates: true
             }
@@ -79,6 +91,7 @@ const handler = async (req: NextApiRequest, res: NextApiResponse) => {
         }
       })
       return res.status(201).json(new_category)
+    }
     default:
       return res.status(405).send('Method Not Allowed')
   }
